Do nothing when there is no recognized name at point.
Quoting Roman:

This is by design. It's only meaningful to talk about the binding site
of a name reference. The name being defined is not a name reference,
it's a name binder. If you do nothing on :no_name (which is what the
vim plugin does, I think), then it shouldn't be confusing.